% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsaux.r
\name{as.matrix.nlsBoot}
\alias{as.matrix.nlsBoot}
\title{Coerce nlsBoot Samples to Matrix}
\usage{
\method{as.matrix}{nlsBoot}(x, ...)
}
\arguments{
\item{x}{An object of class `nlsBoot`.}

\item{...}{Currently ignored.}
}
\value{
a matrix of the bootstrap samples where each row is a
  statistic and each column a sample.
}
\description{
Coerces the samples from nlsBoot to matrix form.
}
\details{
Each row of the result is a statistic and each column a bootstrap
sample.  Samples where the model did not converge are discarded.
}
