% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsaux.r
\name{na_extend}
\alias{na_extend}
\title{Adjust for missing values}
\usage{
na_extend(omit, x)
}
\arguments{
\item{omit}{an object creates by an \code{\link{na.action}} function}

\item{x}{a vector}
}
\value{
a vector
}
\description{
Use missing value information to extend a vector
}
\details{
Cases with missing values are removed from teh fitted values and
weights stored in \code{nls}.  This function reintroduces missing
values at the right locations so that the weights and fitted
values match the original data source.
}
