% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsaux-package.R
\docType{package}
\name{nlsaux-package}
\alias{nlsaux}
\alias{nlsaux-package}
\title{nlsaux: Support functions for nls}
\description{
Provides additional tools for nls.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SWotherspoon/nlsaux}
  \item Report bugs at \url{https://github.com/SWotherspoon/nlsaux/issues}
}

}
\author{
\strong{Maintainer}: Simon Wotherspoon \email{Simon.Wotherspoon@utas.edu.au}

Authors:
\itemize{
  \item Simon Wotherspoon \email{Simon.Wotherspoon@utas.edu.au}
}

}
\keyword{internal}
