% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsaux.r
\name{nlsBayesBoot}
\alias{nlsBayesBoot}
\title{NLS Bayesian Bootstrap}
\usage{
nlsBayesBoot(object, nboot = 99, statistic = coef)
}
\arguments{
\item{object}{A fitted nls object.}

\item{nboot}{Number of bootstrap samples.}

\item{statistic}{A function that takes a fitted nls object as its first
argument and returns a vector of test statistics.}
}
\value{
Returns an object of class `nlsBoot` containing

- `call`: the matched call
- `statistic`: the test statistic
- `nls`: the fitted nls object
- `boot`: an array of bootstrap samples
}
\description{
Perform a Bayesian bootstrap for a fitted nls model.
}
\details{
The user must provide a function to compute the test statistic
from a fitted nls object. The test statistic is computed for each
bootstrap sample and the results stored as a list in the `boot`
component of the returned object. If [nls()] fails to converge
for any bootstrap sample `NULL` is returned for that sample.
}
