% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsaux.r
\name{nlsWildBoot}
\alias{nlsWildBoot}
\title{NLS Wild Bootstrap}
\usage{
nlsWildBoot(object, nboot = 99, statistic = coef)
}
\arguments{
\item{object}{A fitted nls object.}

\item{nboot}{Number of bootstrap samples.}

\item{statistic}{a function that takes a fitted nls object as its first
arguement and returns a vector of test statistics.}
}
\value{
Returns an object of class \code{nlsBoot} containing
\item{\code{call}}{the matched call}
\item{\code{statistic}}{the test statistic}
\item{\code{nls}}{the fitted nls object}
\item{\code{boot}}{an array of bootrap samples}
}
\description{
Perform a wild bootstrap for a fitted nls model.
}
\details{
The user must provide a function to compute the test statistic
from a fitted nls object.  The standard Normal distribution is
used to construct wild bootstrap samples from the fitted model.
The test statistic is computed for each bootstrap sample and the
results stored as a list in the \code{boot} component of the
returned object. If \code{nls} fails to converge for any bootstrap
sample \code{NULL} is returned for that sample.
}
