% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsaux.r
\name{plot.nlsBoot}
\alias{plot.nlsBoot}
\title{Plots Samples from nlsBoot}
\usage{
\method{plot}{nlsBoot}(x, which, ...)
}
\arguments{
\item{x}{An object of class \code{nlsBoot}.}

\item{which}{A numeric vector indicating which statistics to plot.}

\item{...}{Extra arguments passed to the appropriate plot
function.}
}
\description{
Plots of the bootstrap samples of an \code{nlsBoot} object.
}
\details{
If \code{which} identifies a single statistic a box plot is
contructed with a call to \code{boxplot}, for two statistics a
standard plot is constructed with a call to \code{plot} and for
more than two a pairwise plot is constructed with a call to
\code{pairs}.
}
