% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsaux.r
\name{qqnorm.nlsBoot}
\alias{qqnorm.nlsBoot}
\title{Normal QQ Plots for nlsBoot}
\usage{
\method{qqnorm}{nlsBoot}(y, which, ...)
}
\arguments{
\item{y}{an object of class \code{nlsBoot}}

\item{which}{a numeric vector indicating with statistics to plot}

\item{...}{extra arguments to be passed to \code{qqnorm}}
}
\description{
Normal QQ plots of the bootstrap samples of an \code{nlsBoot} object.
}
