% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsaux.r
\name{simulate.nls}
\alias{simulate.nls}
\title{Simulate Responses.}
\usage{
\method{simulate}{nls}(object, nsim = 1, seed = NULL, ...)
}
\arguments{
\item{object}{An object representing a fitted model.}

\item{nsim}{Number of response vectors to simulate. Defaults to 1.}

\item{seed}{An object specifying if and how the random number
generator should be initialized ('seeded').}

\item{...}{Additional optional arguments.}
}
\value{
A dataframe with attribute "seed", where each column
  represents a set of simulated responses.
}
\description{
Simulate one or more responses from the distribution corresponding
to a fitted [nls()] object.
}
\details{
This is a simple wrapper function for [simulate.lm()].
}
