% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsaux.r
\name{summary.nlsBoot}
\alias{summary.nlsBoot}
\alias{print.summary.nlsBoot}
\title{Summary for nlsBoot}
\usage{
\method{summary}{nlsBoot}(object, level = 0.95, method = c("basic",
  "percentile", "Normal"), ...)

\method{print}{summary.nlsBoot}(x, digits = max(3L, getOption("digits") - 3L),
  ...)
}
\arguments{
\item{object}{an object of class \code{nlsBoot}}

\item{level}{the confidence level required.}

\item{method}{confidence interval type}

\item{...}{ignored}

\item{x}{an object of class \code{summary.nlsBoot}}

\item{digits}{the number of significant digits to use when printing}
}
\value{
Returns the estimates as an array.
}
\description{
Constructs a table of estimates and boostrap confidence intervals for
the test statistic.
}
