% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsaux.r
\name{na_extend}
\alias{na_extend}
\title{Adjust for missing values}
\usage{
na_extend(omit, x)
}
\arguments{
\item{omit}{An object created by an \code{\link{na.action}} function}

\item{x}{A vector or matrix}
}
\value{
A vector or matrix
}
\description{
Use missing value information to extend a vector
}
\details{
The data, weights and fitted values stored within an
\code{nlsModel} object have had cases with any missing values
filtered out.  This function reintroduces missing values at the
right locations so that the data, weights and fitted values match
the original data source.
}
