% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsaux.r
\name{nlsJKnife}
\alias{nlsJKnife}
\title{NLS Jackknife}
\usage{
nlsJKnife(object, statistic = coef)
}
\arguments{
\item{object}{a fitted nls object.}

\item{statistic}{a function that takes a fitted nls object as its first
argument and returns a vector of test statistics.}
}
\value{
Returns an object of class \code{nlsJKnife} containing
\item{\code{call}}{the matched call}
\item{\code{statistic}}{the test statistic}
\item{\code{nls}}{the fitted nls object}
\item{\code{jknife}}{an array of jackknife samples}
}
\description{
Perform the Jackknife for a fitted nls model.
}
\details{
The user must provide a function to compute the test statistic
from a fitted nls object.  The test statistic is computed for each
jacknife sample.
}
