% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsaux.r
\name{nlsParBoot}
\alias{nlsParBoot}
\title{NLS Parametric Bootstrap}
\usage{
nlsParBoot(object, nboot = 99, stat = coef)
}
\arguments{
\item{object}{a fitted nls object.}

\item{nboot}{number of bootstrap samples.}

\item{stat}{a function that takes a fitted nls object and returns
the test statistic.}
}
\value{
an array of bootstrap samples.
}
\description{
Perfrom a parametric bootstrap for a fitted nls model.
}
\details{
The user must provide a function to compute the test statistic
from a fitted nls object.  The test statistic is computed for each
bootstrap sample and the results are returned as an array.
}
