% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsaux.r
\name{nlsParBoot}
\alias{nlsParBoot}
\title{NLS Parametric Bootstrap}
\usage{
nlsParBoot(object, nboot = 99, statistic = coef)
}
\arguments{
\item{object}{a fitted nls object.}

\item{nboot}{number of bootstrap samples.}

\item{statistic}{a function that takes a fitted nls object as its first
arguement and returns a vector of test statistics.}
}
\value{
Returns an object of class \code{nlsBoot} containing
\item{\code{call}}{the matched call}
\item{\code{statistic}}{the test statistic}
\item{\code{nls}}{the fitted nls object}
\item{\code{boot}}{an array of bootrap samples}
}
\description{
Perform a parametric bootstrap for a fitted nls model.
}
\details{
The user must provide a function to compute the test statistic
from a fitted nls object.  The test statistic is computed for each
bootstrap sample and the results are returned as an array.
}
