% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsaux.r
\name{qqnorm.nlsBoot}
\alias{qqnorm.nlsBoot}
\title{Normal QQ Plots for nlsBoot}
\usage{
\method{qqnorm}{nlsBoot}(y, which, ...)
}
\arguments{
\item{y}{An object of class `nlsBoot`.}

\item{which}{A numeric vector indicating which statistics to plot.}

\item{...}{Extra arguments to pass to [qqnorm()].}
}
\description{
Normal QQ plots of the bootstrap samples of an `nlsBoot`
object.
}
