% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlsaux.r
\name{summary.nlsJKnife}
\alias{summary.nlsJKnife}
\alias{print.summary.nlsJKnife}
\title{Summary for nlsJKnife}
\usage{
\method{summary}{nlsJKnife}(object, level = 0.95, ...)

\method{print}{summary.nlsJKnife}(x, digits = max(3L, getOption("digits") -
  3L), ...)
}
\arguments{
\item{object}{An object of class \code{nlsJKnife}.}

\item{level}{The confidence level required.}

\item{...}{Currently ignored.}

\item{x}{An object of class \code{summary.nlsJKnife}.}

\item{digits}{The number of significant digits to use when printing.}
}
\value{
Returns an object of class \code{nlsJKnife} containing
\item{\code{object}}{the \code{nlsJKnife} object}
\item{\code{theta}}{estimate of the test statistic}
\item{\code{corrected}}{jackknife bias corrected estimate}
\item{\code{bias}}{jackknife estimate of bias}
\item{\code{cov}}{jackknife estimate of covariance}
\item{\code{level}}{requested confidence level}
\item{\code{lwr}}{jackknife lower confidence limit}
\item{\code{upr}}{jackknife upper confidence limit}
}
\description{
Constructs a table of estimates and confidence intervals for
the test statistic.
}
